add --tls-disable flag

+++ test/io.mock.c
@@ -35,7 +35,7 @@ io_sendf(struct connection *c, const char *fmt, ...)
 }
 
 struct connection*
-connection(const void *o, const char *h, const char *p, uint8_t f)
+connection(const void *o, const char *h, const char *p, uint32_t f)
 {
 	UNUSED(o);
 	UNUSED(h);
